+++ fairscale/experimental/nn/distributed_pipeline/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) Facebook, Inc. and its affiliates.
+#
+# This source code is licensed under the BSD license found in the
+# LICENSE file in the root directory of this source tree.
+
+from .graph import PipelineModulesGraph
+from .loss import DistributedLoss
+from .pipeline import DistributedPipeline
